move api calls to helper file, rename helper functions
