+++ frontend/src/components/Modal/EditModal.jsx
@@ -16,7 +16,7 @@ import {
 import { useDisclosure } from "@chakra-ui/react";
 import { employeesContext } from "../../Providers/EmployeesProvider";
 import { modalContext } from "../../Providers/ModalProvider";
-import { onUpdateHandler } from "../../helpers/helpers";
+import { updateEmployeeHandler } from "../../helpers/helpers";
 
 const EditModal = ({ id }) => {
     const { employees, setEmployees } = useContext(employeesContext);
@@ -50,37 +50,6 @@ const EditModal = ({ id }) => {
             salary,
         });
     }
-
-    // async function onUpdateHandler(id) {
-    //     const employeeId = id;
-    //     const employeeData = editFormValues;
-
-    //     try {
-    //         const url = `/api/employees/${employeeId}`;
-    //         const response = await fetch(url, {
-    //             method: "PUT",
-    //             headers: {
-    //                 "Content-type": "application/json",
-    //             },
-    //             body: JSON.stringify(employeeData),
-    //         });
-    //         const data = await response.json();
-    //         const updatedEmployee = data.results.rows;
-
-    //         const newEmployeeList = employees.map(
-    //             (employee) =>
-    //                 updatedEmployee.find(
-    //                     (updated) => updated.id === employee.id
-    //                 ) || employee
-    //         );
-
-    //         setEmployees(newEmployeeList);
-    //         setModalState("None");
-    //         onClose();
-    //     } catch (err) {
-    //         console.log(err);
-    //     }
-    // }
 
     return (
         <>
@@ -155,7 +124,7 @@ const EditModal = ({ id }) => {
                             colorScheme="teal"
                             size="md"
                             onClick={() => {
-                                onUpdateHandler(
+                                updateEmployeeHandler(
                                     editFormValues.id,
                                     editFormValues,
                                     employees,
